linkedin-stats-automation/requirements.txt
functions-framework==3.5.0
google-cloud-secret-manager==2.16.3
google-cloud-storage==2.10.0
google-cloud-logging==3.5.0
gspread==5.10.0
oauth2client==4.1.3
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
urllib3==2.0.7
python-dotenv==1.0.0